Export results to output file
- Added function file_output to write analysis results to an external file
- Small changes to I/O validation feedback for clarity
- Corrected variance calculation to calculate sample variance rather than population variance
- Small formatting changes made across project to improve consistency and readability
- Commented out test function report_clipping

+++ waveform.h
@@ -28,7 +28,7 @@ void compute_rms(double *rms_A, double *rms_B, double *rms_C);
 void compute_p2p(double *p2p_A, double *p2p_B, double *p2p_C);
 void compute_dc_offset(double *mean_A, double *mean_B, double *mean_C);
 void detect_clipping();
-void check_rms_tolerance(double *tolerance_status, int index, double* rms_value);
+void check_rms_tolerance(int *tolerance_status, int index, double* rms_value);
 void compute_variance_std_dev(double *var_A, double *var_B, double *var_C, double *stddev_A, double *stddev_B, double *stddev_C);
 
 #endif //PQ_ANALYSER_WAVEFORM_H
